Refactor node configuration handling
